add delete landing banner controller

+++ controllers/category.controller.js
@@ -2,20 +2,23 @@ import Category from "../models/category.model.js";
 import { deleteFile } from "../utils/delete-file.js";
 
 export async function addCategory(req, res) {
-  console.log(req?.body, req?.file);
-
-  if (!req?.body?.name || !req?.body?.description || !req?.file?.filename) {
-    deleteFile(req?.file?.filename);
-    return res.status(400).json({ message: "One of more fields are missing" });
+  const { filename } = req?.file;
+  const { name, description } = req?.body;
+  if (!name || !description || !filename) {
+    if (filename) deleteFile(filename);
+    return res.status(400).json({
+      message:
+        "required fields: name, description, icon, One of more fields are missing",
+    });
   }
   try {
     const category = await Category.create({
-      icon_url: req?.file?.filename,
-      name: req?.body?.name,
-      description: req?.body?.description,
+      icon_url: filename,
+      name,
+      description,
     });
 
-    return res.status(201).json({ category });
+    return res.status(201).json({ status: "success", category });
   } catch (err) {
     console.log(err);
     res.status(500).json({ message: "internal server error" });
@@ -33,7 +36,7 @@ export async function getAllCategories(req, res) {
       }
     }
     const categories = await Category.find(query);
-    return res.status(200).json({ categories });
+    return res.status(200).json({ status: "success", categories });
   } catch (err) {
     console.log(err);
     res.status(500).json({ message: "internal server error" });
@@ -53,7 +56,7 @@ export async function getCategoryById(req, res) {
         .json({ message: "invalid categoryId, category not found" });
     }
 
-    return res.status(200).json({ category });
+    return res.status(200).json({ status: "success", category });
   } catch (err) {
     console.log(err);
     res.status(500).json({ message: "internal server error" });
@@ -63,7 +66,7 @@ export async function getCategoryById(req, res) {
 export async function updateCatgory(req, res) {
   try {
     if (!req?.body?.categoryId) {
-      deleteFile(req?.file?.filename);
+      if (req?.file?.filename) deleteFile(req?.file?.filename);
       return res.status(400).json({ message: "categoryId is missing" });
     }
     const categoryBody = {};
@@ -82,6 +85,12 @@ export async function updateCatgory(req, res) {
     }
 
     const catgory = await Category.findById(req?.body?.categoryId);
+    if (!catgory) {
+      if (req?.file?.filename) deleteFile(req?.file?.filename);
+      return res
+        .status(404)
+        .json({ message: "invalid categoryId, category not found" });
+    }
 
     const updatedCategory = await Category.findByIdAndUpdate(
       req?.body?.categoryId,
@@ -90,7 +99,9 @@ export async function updateCatgory(req, res) {
     );
 
     deleteFile(catgory?.icon_url);
-    return res.status(200).json({ category: updatedCategory });
+    return res
+      .status(200)
+      .json({ status: "success", category: updatedCategory });
   } catch (err) {
     console.log(err);
     res.status(500).json({ message: "internal server error" });
@@ -104,10 +115,16 @@ export async function deleteCategory(req, res) {
       return res.status(400).json({ message: "categoryId is missing" });
     }
     const deletedCategory = await Category.findByIdAndDelete(categoryId);
-    // fs.unlinkSync(`${uploadPath}/${deletedCategory?.icon_url}`);
+    if (!deletedCategory) {
+      return res
+        .status(404)
+        .json({ message: "invalid categoryId, category not found" });
+    }
     deleteFile(deletedCategory?.icon_url);
 
-    return res.status(200).json({ status: "success" });
+    return res
+      .status(200)
+      .json({ status: "success", message: "category deleted successfully" });
   } catch (err) {
     console.log(err);
     res.status(500).json({ message: "internal server error" });
